fase_2/Evidencias Proyecto/Evidencias de sistema Aplicación/nfc_ignite_web/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
arrow==1.3.0
binaryornot==0.4.4
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cookiecutter==2.6.0
cryptography==43.0.3
fastapi==0.115.5
flet==0.24.1
flet-core==0.24.1
flet-runtime==0.24.1
gcloud==0.18.3
googleapis-common-protos==1.66.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
jwcrypto==1.5.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
oauth2client==4.1.3
oauthlib==3.2.2
packaging==23.2
protobuf==5.29.0rc3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.2.0
pypng==0.20220715.0
Pyrebase4==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jwt==4.1.0
python-slugify==8.0.4
PyYAML==6.0.2
qrcode==7.4.2
repath==0.9.0
requests==2.32.3
requests-toolbelt==0.10.1
rich==13.9.4
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.41.3
text-unidecode==1.3
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
urllib3==1.26.20
uvicorn==0.32.0
watchdog==4.0.2
watchfiles==0.24.0
websockets==14.1
